backend/requirements.txt
# Image-to-Video Generator - Python Dependencies
# Python 3.10 required

# ============================================================================
# PyTorch with CUDA 11.8 Support
# ============================================================================
# Install these first with special index URL:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# ============================================================================
# Diffusers and Transformers (Core AI Models)
# ============================================================================
diffusers==0.24.0          # SVD and AnimateDiff pipelines
transformers==4.36.0        # CLIP, text encoders
accelerate==0.25.0          # Memory optimization
safetensors==0.4.1          # Safe model format

# ============================================================================
# Memory Optimization
# ============================================================================
xformers==0.0.23            # Memory-efficient attention (CRITICAL for 12GB VRAM)

# ============================================================================
# Image and Video Processing
# ============================================================================
opencv-python==4.8.1.78     # Video encoding/decoding
Pillow==10.1.0              # Image processing
imageio==2.33.0             # Image I/O
imageio-ffmpeg==0.4.9       # Video I/O backend

# ============================================================================
# Model Download and Management
# ============================================================================
huggingface-hub==0.19.4     # Download models from HuggingFace
requests==2.31.0            # HTTP requests for CivitAI downloads
tqdm==4.66.1                # Progress bars for downloads

# ============================================================================
# Utilities
# ============================================================================
numpy==1.24.3               # Numerical operations (compatible with torch)
scipy==1.11.4               # Scientific computing
omegaconf==2.3.0            # Configuration management

# ============================================================================
# Optional: Development and Testing
# ============================================================================
pytest==7.4.3               # Testing framework
black==23.12.1              # Code formatting
flake8==6.1.0               # Linting

# ============================================================================
# Installation Instructions
# ============================================================================
#
# IMPORTANT: Install PyTorch with CUDA support FIRST:
#
# Step 1: Install PyTorch with CUDA 11.8
# pip install torch==2.1.0 torchvision==0.16.0 torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cu118
#
# Step 2: Install all other dependencies
# pip install -r requirements.txt
#
# OR use the automated setup script:
# bash scripts/setup_wsl.sh
#
# ============================================================================
# Version Notes
# ============================================================================
#
# PyTorch 2.1.0:
#   - CUDA 11.8 support
#   - Stable for production
#   - Compatible with xformers 0.0.23
#
# Diffusers 0.24.0:
#   - SVD-XT support
#   - AnimateDiff support
#   - Stable API
#
# Xformers 0.0.23:
#   - CRITICAL for 12GB VRAM
#   - Reduces memory usage by 30-40%
#   - May need specific CUDA version
#
# Python 3.10:
#   - Required for xformers compatibility
#   - Stable with all dependencies
#
# ============================================================================
# Troubleshooting
# ============================================================================
#
# If xformers installation fails:
#   1. Ensure CUDA 11.8 is installed
#   2. Try: pip install xformers --no-deps
#   3. Or install from source (takes time)
#
# If torch doesn't see CUDA:
#   1. Check nvidia-smi works
#   2. Reinstall torch with correct CUDA version
#   3. Verify with: python -c "import torch; print(torch.cuda.is_available())"
#
# If out of memory during installation:
#   1. Close other applications
#   2. Install packages one by one
#   3. Use --no-cache-dir flag: pip install --no-cache-dir -r requirements.txt
#
# ============================================================================
